add base stmt

+++ models.py
@@ -7,7 +7,7 @@ Base = declarative_base()
 class Users(Base):
     __tablename__ = 'users'
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
     name = Column(String(255), nullable=False, unique=True, index=True)
     email = Column(String(255), nullable=False, index=True)
     password = Column(String(255), nullable=False)
@@ -16,7 +16,7 @@ class Users(Base):
 class Routes(Base):
     __tablename__ = 'routes'
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
     name = Column(String(255), nullable=False, unique=True, index=True)
     points = Column(ARRAY(String(200)), default=list())
 
@@ -24,7 +24,7 @@ class Routes(Base):
 class Points(Base):
     __tablename__ = 'points'
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
     name = Column(String(255), nullable=False, unique=True, index=True)
     childs = Column(ARRAY(String(200)), default=list())
 
@@ -32,7 +32,7 @@ class Points(Base):
 class UserRoutes(Base):
     __tablename__ = 'user_routes'
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
     route_name = Column(String(255), nullable=False, unique=True, index=True)
     passed_routes = Column(ARRAY(String(200)), default=list())
-    user_id = Column(BigInteger, ForeignKey(f'{Users.__tablename__}.id'), index=True)
+    user_id = Column(Integer, ForeignKey(f'{Users.__tablename__}.id'), index=True)
